fix(f16): accumulate the scalar 1xN and packed conv1d f16 legs in float32
Apply the #449 rule (accumulate bias and products in float32, round to
float16 once) to scalar float16 legs that do not share an accumulator with an
MVE float16-lane path:

- arm_convolve_1_x_n_f16: the non-MVE dot returns a float32 sum and the
  remainder loop adds the bias in float32 before rounding and clamping, as
  #462 did for arm_nn_mat_mult_nt_t_f16.
- arm_nn_conv1d_k3_packed_f16 and arm_nn_conv1d_k5_packed_f16: the non-MVE
  loops accumulate in float32 and round once at the store.

The MVE legs of these kernels are unchanged, and their generated code on MVE
builds is identical. The arm_convolve_f16 note and the packed conv1d helper
docs state the accumulation width per leg.

test_arm_convolve_packed_f16 gains long-chain cases on the scalar leg for the
k=3 and k=5 packed specializations and the 1xN no-padding region: an all-ones
chain just over 2048 taps must give the exact tap count, and a non-integer
chain must be within one float16 ulp of the float32 reference. Reverting any
one site fails it on GCC 13.2 and 14.3. The MVE leg reports it as ignored. Left for separate changes:
- Sites where an MVE float16-lane path can reach the same layer in the same
  build: batch_matmul, LSTM and GRU, and the depthwise ch_mult > 1 generic
  kernel, which shares one-input-channel layers with the to-convolution route.
- The scalar remainders inside the MVE blocks of the non-packed conv1d
  kernels.
- The SVDF scalar leg, which is not in #465's list.

Refs #465
Refs #386

+++ Source/NNSupportFunctions/arm_nn_conv1d_k5_packed_f16.c
@@ -175,16 +175,17 @@ void arm_nn_conv1d_k5_packed_f16(const float16_t *__RESTRICT x_nhwc,
         {
             const int32_t lane = oc % block_cols;
             const float16_t *w_base = kernel_packed + ((size_t)oc / block_cols) * 5U * (size_t)in_c * block_cols;
-            _Float16 acc = b ? (_Float16)b[oc] : (_Float16)0.0f;
+            /* Accumulate in float32 and round to f16 once at the store (#449, #465). */
+            float32_t acc = b ? (float32_t)b[oc] : 0.0f;
 
             for (int32_t ic = 0; ic < in_c; ++ic)
             {
                 const float16_t *w_ic = w_base + (size_t)ic * block_cols;
-                acc += (_Float16)x0[ic] * (_Float16)w_ic[(size_t)0 * in_c * block_cols + lane];
-                acc += (_Float16)x1[ic] * (_Float16)w_ic[(size_t)1 * in_c * block_cols + lane];
-                acc += (_Float16)x2[ic] * (_Float16)w_ic[(size_t)2 * in_c * block_cols + lane];
-                acc += (_Float16)x3[ic] * (_Float16)w_ic[(size_t)3 * in_c * block_cols + lane];
-                acc += (_Float16)x4[ic] * (_Float16)w_ic[(size_t)4 * in_c * block_cols + lane];
+                acc += (float32_t)x0[ic] * (float32_t)w_ic[(size_t)0 * in_c * block_cols + lane];
+                acc += (float32_t)x1[ic] * (float32_t)w_ic[(size_t)1 * in_c * block_cols + lane];
+                acc += (float32_t)x2[ic] * (float32_t)w_ic[(size_t)2 * in_c * block_cols + lane];
+                acc += (float32_t)x3[ic] * (float32_t)w_ic[(size_t)3 * in_c * block_cols + lane];
+                acc += (float32_t)x4[ic] * (float32_t)w_ic[(size_t)4 * in_c * block_cols + lane];
             }
 
             y[oc] = (float16_t)acc;
